Fix ambiguous cos/sin call in SkillRadarView
radius is CGFloat (from CGRect geometry) and angle is Double, so the
compiler couldn't choose between cos(Double) and cos(CGFloat). Cast
angle to CGFloat so all operands are consistently CGFloat.

+++ Chess4U/Views/Dashboard/SkillRadarView.swift
@@ -110,8 +110,8 @@ struct RadarShape: Shape {
         for (i, value) in values.enumerated() {
             let angle = Double(i) * 2 * .pi / Double(values.count) - .pi / 2
             let r = radius * max(0.05, value)
-            let point = CGPoint(x: center.x + r * cos(angle),
-                                y: center.y + r * sin(angle))
+            let point = CGPoint(x: center.x + r * cos(CGFloat(angle)),
+                                y: center.y + r * sin(CGFloat(angle)))
             if i == 0 { path.move(to: point) }
             else { path.addLine(to: point) }
         }
